+++ DotnetPriceParser/Price.cs
@@ -4,10 +4,10 @@ namespace DotnetPriceParser
 {
     public class Price
     {
-        public float? Amount { get; set; }
+        public double? Amount { get; set; }
         public string Currency { get; set; }
 
-        private Price(float? amount, string currency)
+        private Price(double? amount, string currency)
         {
             Amount = amount;
             Currency = currency;
@@ -15,7 +15,7 @@ namespace DotnetPriceParser
 
         public static Price FromString(string rawPrice)
         {
-            float? amount = AmountParser.parse(rawPrice);
+            double? amount = AmountParser.parse(rawPrice);
             string currency = CurrencyParser.parse(rawPrice);
 
             return new Price(amount, currency);
